관리자 - 부품사 SQ정보 score 타입 Integer -> Double 변경(부품사관리, 상생사업관리, 컨설팅사업관리 dto 수정) - 방문교육 SQ정보 오류 수정

+++ service/src/main/java/com/kap/core/dto/cb/cbb/CBBManageConsultInsertDTO.java
@@ -143,10 +143,10 @@ public class CBBManageConsultInsertDTO extends BaseDTO {
     private String  cbsnSeq;
 
     @Schema(title = "SQ 점수")
-    private Integer  score;
-    private Integer  score1;
-    private Integer  score2;
-    private Integer  score3;
+    private Double score;
+    private Double score1;
+    private Double score2;
+    private Double score3;
 
     @Schema(title = "SQ 평가년도")
     private Integer year;
